- Fix exactly once tests
