+++ src/JustSaying/Messaging/Middleware/Handle/HandlerMiddlewareBuilderExtensions.cs
@@ -27,7 +27,7 @@ namespace JustSaying.Messaging.Middleware.Handle
         {
             HandleMessageMiddleware CreateMiddleware()
             {
-                var messageLock = builder.ServicesBuilder.MessageLock?.Invoke() ?? builder.ServiceResolver.ResolveService<IMessageLockAsync>();
+                var messageLock = builder.ServicesBuilder?.MessageLock?.Invoke() ?? builder.ServiceResolver.ResolveService<IMessageLockAsync>();
                 var logger = builder.ServiceResolver.ResolveService<ILogger<ExactlyOnceMiddleware<TMessage>>>();
 
                 return new ExactlyOnceMiddleware<TMessage>(messageLock,
