+++ tests/JustSaying.IntegrationTests/Fluent/Subscribing/WhenTwoDifferentHandlersHandleAnExactlyOnceMessage.cs
@@ -1,6 +1,8 @@
 using System.Threading.Tasks;
 using JustSaying.IntegrationTests.TestHandlers;
 using JustSaying.Messaging;
+using JustSaying.Messaging.MessageHandling;
+using JustSaying.Messaging.Middleware.Handle;
 using JustSaying.TestingFramework;
 using Microsoft.Extensions.DependencyInjection;
 using Shouldly;
@@ -18,12 +20,16 @@ namespace JustSaying.IntegrationTests.Fluent.Subscribing
         [AwsFact]
         public async Task Then_Both_Handlers_Receive_The_Message()
         {
-            // Arrange
-            var handler1 = new ExactlyOnceHandlerNoTimeout();
-            var handler2 = new ExactlyOnceHandlerNoTimeout();
+            var messageLock = new MessageLockStore();
+
+            var handler1 = new InspectableHandler<SimpleMessage>();
+            var handler2 = new InspectableHandler<SimpleMessage>();
 
             var services = GivenJustSaying()
-                .ConfigureJustSaying((builder) => builder.WithLoopbackTopic<SimpleMessage>(UniqueName))
+                .ConfigureJustSaying(c => c.Services(s => s.WithMessageLock(() => messageLock)))
+                .ConfigureJustSaying((builder) =>
+                    builder.WithLoopbackTopic<SimpleMessage>(UniqueName,
+                        t => t.WithMiddlewareConfiguration(m => m.UseExactlyOnce<SimpleMessage>("some-key"))))
                 .AddJustSayingHandlers(new[] { handler1, handler2 });
 
             await WhenAsync(
@@ -40,8 +46,8 @@ namespace JustSaying.IntegrationTests.Fluent.Subscribing
                     await Task.Delay(1.Seconds(), cancellationToken);
 
                     // Assert
-                    handler1.NumberOfTimesIHaveBeenCalledForMessage(message.UniqueKey()).ShouldBe(1);
-                    handler2.NumberOfTimesIHaveBeenCalledForMessage(message.UniqueKey()).ShouldBe(1);
+                    handler1.ReceivedMessages.ShouldHaveSingleItem().Id.ShouldBe(message.Id);
+                    handler2.ReceivedMessages.ShouldHaveSingleItem().Id.ShouldBe(message.Id);
                 });
         }
     }
