--- tests/JustSaying.IntegrationTests/TestHandlers/ExactlyOnceHandlerNoTimeout.cs
@@ -1,9 +1,0 @@
-using JustSaying.Messaging.MessageHandling;
-
-namespace JustSaying.IntegrationTests.TestHandlers
-{
-    [ExactlyOnce]
-    public class ExactlyOnceHandlerNoTimeout : ExactlyOnceHandlerWithTimeout
-    {
-    }
-}
